# improve PortOut thread safety