--- a/core/modules/port_out.h
+++ b/core/modules/port_out.h
@@ -34,6 +34,7 @@
 #include "../module.h"
 #include "../pb/module_msg.pb.h"
 #include "../port.h"
+#include "../utils/mcslock.h"
 #include "../worker.h"
 
 class PortOut final : public Module {
@@ -41,7 +42,10 @@ class PortOut final : public Module {
   static const gate_idx_t kNumIGates = MAX_GATES;
   static const gate_idx_t kNumOGates = 0;
 
-  PortOut() : Module(), port_(), available_queues_(), worker_queues_() {}
+  PortOut()
+      : Module(), port_(), worker_queues_(), queue_users_(), queue_locks_() {
+    max_allowed_workers_ = Worker::kMaxWorkers;
+  }
 
   CommandResponse Init(const bess::pb::PortOutArg &arg);
 
@@ -56,9 +60,12 @@ class PortOut final : public Module {
  private:
   Port *port_;
 
-  std::vector<queue_t> available_queues_;
-
   int worker_queues_[Worker::kMaxWorkers];
+
+  // Number of workers mapped to a given queue. Indexed by queue number
+  int queue_users_[MAX_QUEUES_PER_DIR];
+
+  mcslock_t queue_locks_[MAX_QUEUES_PER_DIR];
 };
 
 #endif  // BESS_MODULES_PORTOUT_H_