@chris3torek 2019-08-30
2 cycles per packet is about an 11% penalty, but overall I guess that's OK.  The code itself seems fine. Let's see if @sangjinhan has any tricks up his sleeves for more go-fast 😺