@sangjinhan 2019-08-30
Looks good to me overall. A few minor comments:
* IIRC Intel X710 does not support loopback mode, thus there would be no return (RX) traffic. For more accurate measurement of TX overhead, probably using (modified version of) `s2p.bess`
* For common case optimization, where #TXQ >= #workers, it would be ideal bypass spinlocks completely. Even if there are more workers than queues, it might be still possible for some workers to skip locking (e.g., 2 queues & 3 workers, suppose Q0 is shared by W0,W2 and Q1 is used by W1 only (thus no spinlock).
  * This is easily doable only with wid-baased queue selection. So, Q: Is there particular use case of igate-based queue selection you have in mind?
* Please include your BESS used for measurement? (loopback.bess in this case?)