@melvinw 2019-08-30
Thanks @sangjinhan! Cleaned up and checked in the benchmark script. New results below.

# Before

```
$ BESS_WORKERS=1 BESS_QUEUES=1 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source0:0

12:36:12.558757          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W0 !leaf_source0:0      2100.011     3898500     124.752   83833.348      32.000      16.833
--------------------------------------------------------------------------------------------

12:36:13.560111          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W0 !leaf_source0:0      2100.010     3905207     124.967   83977.581      32.000      16.805
--------------------------------------------------------------------------------------------

12:36:14.561424          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W0 !leaf_source0:0      2100.012     3901830     124.859   83904.959      32.000      16.819
--------------------------------------------------------------------------------------------
```

# After (w/o contention)

```
$ BESS_WORKERS=1 BESS_QUEUES=1 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source0:0

13:44:14.300635          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W0 !leaf_source0:0      2100.011     3818578     122.195   82114.722      32.000      17.186
--------------------------------------------------------------------------------------------

13:44:15.302004          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W0 !leaf_source0:0      2100.010     3814724     122.071   82031.839      32.000      17.203
--------------------------------------------------------------------------------------------

13:44:16.303339          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W0 !leaf_source0:0      2100.011     3811088     121.955   81953.656      32.000      17.220
--------------------------------------------------------------------------------------------
```

```
$ BESS_WORKERS=2 BESS_QUEUES=2 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source1:0, !leaf_source0:0

13:44:57.834161          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W1 !leaf_source1:0      2100.011     3641251     116.520   78301.466      32.000      18.023
W0 !leaf_source0:0      2100.011     3643326     116.586   78346.100      32.000      18.012
--------------------------------------------------------------------------------------------

13:44:58.835741          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W1 !leaf_source1:0      2100.009     3647734     116.727   78440.878      32.000      17.991
W0 !leaf_source0:0      2100.010     3646166     116.677   78407.155      32.000      17.998
--------------------------------------------------------------------------------------------

13:44:59.837272          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W1 !leaf_source1:0      2100.012     3647205     116.711   78429.498      32.000      17.993
W0 !leaf_source0:0      2100.010     3647170     116.709   78428.746      32.000      17.993
--------------------------------------------------------------------------------------------
```

```
$ BESS_WORKERS=4 BESS_QUEUES=4 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source3:0, !leaf_source2:0, !leaf_source0:0, !leaf_source1:0

13:45:17.452483          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W3 !leaf_source3:0      2100.010     3358374     107.468   72218.494      32.000      19.541
W2 !leaf_source2:0      2100.011     3096680      99.094   66591.008      32.000      21.192
W0 !leaf_source0:0      2100.012     4181622     133.812   89921.612      32.000      15.694
W1 !leaf_source1:0      2100.010     3804449     121.742   81810.879      32.000      17.250
--------------------------------------------------------------------------------------------

13:45:18.454504          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W3 !leaf_source3:0      2100.011     3351086     107.235   72061.772      32.000      19.583
W2 !leaf_source2:0      2100.009     3100655      99.221   66676.506      32.000      21.165
W0 !leaf_source0:0      2100.010     4180364     133.772   89894.551      32.000      15.698
W1 !leaf_source1:0      2100.010     3805486     121.776   81833.181      32.000      17.245
--------------------------------------------------------------------------------------------

13:45:19.456435          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W3 !leaf_source3:0      2100.012     3353832     107.323   72120.815      32.000      19.567
W2 !leaf_source2:0      2100.012     3104156      99.333   66751.787      32.000      21.141
W0 !leaf_source0:0      2100.011     4182393     133.837   89938.181      32.000      15.691
W1 !leaf_source1:0      2100.013     3804978     121.759   81822.250      32.000      17.247
--------------------------------------------------------------------------------------------
```

```
$ BESS_WORKERS=6 BESS_QUEUES=6 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source5:0, !leaf_source4:0, !leaf_source3:0, !leaf_source2:0, !leaf_source0:0, !leaf_source1:0

13:46:55.167737          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W5 !leaf_source5:0      2100.013     3782918     121.053   81347.883      32.000      17.348
W4 !leaf_source4:0      2100.010     3347537     107.121   71985.438      32.000      19.604
W3 !leaf_source3:0      2100.011     3309054     105.890   71157.918      32.000      19.832
W2 !leaf_source2:0      2100.011     3538082     113.219   76082.922      32.000      18.548
W0 !leaf_source0:0      2100.011     3732933     119.454   80273.004      32.000      17.580
W1 !leaf_source1:0      2100.011     3505540     112.177   75383.152      32.000      18.720
--------------------------------------------------------------------------------------------

13:46:56.170233          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W5 !leaf_source5:0      2100.010     3782738     121.048   81344.006      32.000      17.349
W4 !leaf_source4:0      2100.009     3347525     107.121   71985.178      32.000      19.604
W3 !leaf_source3:0      2100.011     3304511     105.744   71060.216      32.000      19.859
W2 !leaf_source2:0      2100.010     3532386     113.036   75960.445      32.000      18.578
W0 !leaf_source0:0      2100.010     3732662     119.445   80267.169      32.000      17.581
W1 !leaf_source1:0      2100.009     3504375     112.140   75358.081      32.000      18.727
--------------------------------------------------------------------------------------------

13:46:57.172687          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W5 !leaf_source5:0      2100.009     3779999     120.960   81285.099      32.000      17.361
W4 !leaf_source4:0      2100.011     3348648     107.157   72009.338      32.000      19.598
W3 !leaf_source3:0      2100.012     3302642     105.685   71020.031      32.000      19.871
W2 !leaf_source2:0      2100.010     3536845     113.179   76056.325      32.000      18.555
W0 !leaf_source0:0      2100.009     3730429     119.374   80219.163      32.000      17.592
W1 !leaf_source1:0      2100.012     3503387     112.108   75336.846      32.000      18.732
--------------------------------------------------------------------------------------------
```

# After (w/ contention)

```
$ BESS_WORKERS=2 BESS_QUEUES=1 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source1:0, !leaf_source0:0

13:58:35.653585          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W1 !leaf_source1:0      2100.009     1340777      42.905   28832.074      32.000      48.946
W0 !leaf_source0:0      2100.010     1357836      43.451   29198.922      32.000      48.331
--------------------------------------------------------------------------------------------

13:58:36.655191          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W1 !leaf_source1:0      2100.011     1350957      43.231   29050.992      32.000      48.577
W0 !leaf_source0:0      2100.011     1367842      43.771   29414.080      32.000      47.977
--------------------------------------------------------------------------------------------

13:58:37.656740          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W1 !leaf_source1:0      2100.012     1349581      43.187   29021.402      32.000      48.626
W0 !leaf_source0:0      2100.011     1366735      43.736   29390.289      32.000      48.016
--------------------------------------------------------------------------------------------
```

```
$ BESS_WORKERS=4 BESS_QUEUES=1 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source3:0, !leaf_source2:0, !leaf_source0:0, !leaf_source1:0

13:58:48.559053          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W3 !leaf_source3:0      2100.011      740813      23.706   15930.448      32.000      88.586
W2 !leaf_source2:0      2100.008      741191      23.718   15938.575      32.000      88.540
W0 !leaf_source0:0      2100.009      737840      23.611   15866.522      32.000      88.942
W1 !leaf_source1:0      2100.012     3182243     101.832   68430.959      32.000      20.622
--------------------------------------------------------------------------------------------

13:58:49.561037          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W3 !leaf_source3:0      2100.010      743583      23.795   15990.020      32.000      88.255
W2 !leaf_source2:0      2100.015      743801      23.802   15994.711      32.000      88.230
W0 !leaf_source0:0      2100.013      740824      23.706   15930.683      32.000      88.584
W1 !leaf_source1:0      2100.009     3183345     101.867   68454.670      32.000      20.615
--------------------------------------------------------------------------------------------

13:58:50.562962          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W3 !leaf_source3:0      2100.013      743481      23.791   15987.830      32.000      88.268
W2 !leaf_source2:0      2100.008      743834      23.803   15995.424      32.000      88.226
W0 !leaf_source0:0      2100.011      741256      23.720   15939.974      32.000      88.533
W1 !leaf_source1:0      2100.010     3182896     101.853   68445.016      32.000      20.618
--------------------------------------------------------------------------------------------
```

```
$ BESS_WORKERS=6 BESS_QUEUES=1 bin/bessctl daemon start -- run port/s2p -- monitor tc
Environment variable "BESS_PORTS" is not set. Using default value "1"
Using 1 DPDK ports... (envvar "BESS_PORTS")
Monitoring traffic classes: !leaf_source5:0, !leaf_source4:0, !leaf_source3:0, !leaf_source2:0, !leaf_source0:0, !leaf_source1:0

14:01:18.896122          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W5 !leaf_source5:0      2100.011     2735193      87.526   58817.596      32.000      23.993
W4 !leaf_source4:0      2100.012      603359      19.308   12974.646      32.000     108.767
W3 !leaf_source3:0      2100.011      584015      18.688   12558.659      32.000     112.369
W2 !leaf_source2:0      2100.012      582186      18.630   12519.332      32.000     112.722
W0 !leaf_source0:0      2100.010     2746449      87.886   59059.641      32.000      23.895
W1 !leaf_source1:0      2100.018      601206      19.239   12928.345      32.000     109.156
--------------------------------------------------------------------------------------------

14:01:19.898598          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W5 !leaf_source5:0      2100.009     2726435      87.246   58629.277      32.000      24.070
W4 !leaf_source4:0      2100.010      606378      19.404   13039.568      32.000     108.225
W3 !leaf_source3:0      2100.012      585806      18.746   12597.176      32.000     112.026
W2 !leaf_source2:0      2100.012      583324      18.666   12543.818      32.000     112.502
W0 !leaf_source0:0      2100.010     2719845      87.035   58487.562      32.000      24.128
W1 !leaf_source1:0      2100.008      602189      19.270   12949.494      32.000     108.978
--------------------------------------------------------------------------------------------

14:01:20.901010          CPU MHz   scheduled        Mpps        Mbps  pkts/sched    cycles/p
--------------------------------------------------------------------------------------------
W5 !leaf_source5:0      2100.012     2740024      87.681   58921.486      32.000      23.951
W4 !leaf_source4:0      2100.010      607428      19.438   13062.148      32.000     108.038
W3 !leaf_source3:0      2100.011      586703      18.774   12616.462      32.000     111.854
W2 !leaf_source2:0      2100.009      584121      18.692   12560.949      32.000     112.349
W0 !leaf_source0:0      2100.012     2759309      88.298   59336.194      32.000      23.783
W1 !leaf_source1:0      2100.013      602332      19.275   12952.562      32.000     108.952
--------------------------------------------------------------------------------------------
```